START: Refactor PSM and PS
