+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/state/PartitionStateManager.java
@@ -22,8 +22,6 @@ import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 
-import static io.confluent.csid.utils.JavaUtils.getFirst;
-import static io.confluent.csid.utils.JavaUtils.getLast;
 import static io.confluent.csid.utils.KafkaUtils.toTopicPartition;
 
 /**
@@ -220,40 +218,14 @@ public class PartitionStateManager<K, V> implements ConsumerRebalanceListener {
         }
     }
 
-    /**
-     * Have our partitions been revoked?
-     * <p>
-     * This state is rare, as shards or work get removed upon partition revocation, although under busy load it might
-     * occur we don't synchronize over PartitionState here so it's a bit racey, but is handled and eventually settles.
-     *
-     * @return true if epoch doesn't match, false if ok
-     */
-    boolean checkIfWorkIsStale(final WorkContainer<?, ?> workContainer) {
-        var topicPartitionKey = workContainer.getTopicPartition();
-
-        Long currentPartitionEpoch = partitionsAssignmentEpochs.get(topicPartitionKey);
-        long workEpoch = workContainer.getEpoch();
-
-        boolean partitionNotAssigned = isPartitionRemovedOrNeverAssigned(workContainer.getCr());
-
-        boolean epochMissMatch = currentPartitionEpoch != workEpoch;
-
-        if (epochMissMatch || partitionNotAssigned) {
-            log.debug("Epoch mismatch {} vs {} for record {}. Skipping message - it's partition has already assigned to a different consumer.",
-                    workEpoch, currentPartitionEpoch, workContainer);
-            return true;
-        }
-        return false;
-    }
-
-    // todo move to partition state
-    public boolean isRecordPreviouslyCompleted(ConsumerRecord<K, V> rec) {
-        var tp = toTopicPartition(rec);
-        var partitionState = getPartitionState(tp);
-        boolean previouslyCompleted = partitionState.isRecordPreviouslyCompleted(rec);
-        log.trace("Record {} previously completed? {}", rec.offset(), previouslyCompleted);
-        return previouslyCompleted;
-    }
+//    // todo move to partition state
+//    public boolean isRecordPreviouslyCompleted(ConsumerRecord<K, V> rec) {
+//        var tp = toTopicPartition(rec);
+//        var partitionState = getPartitionState(tp);
+//        boolean previouslyCompleted = partitionState.isRecordPreviouslyCompleted(rec);
+//        log.trace("Record {} previously completed? {}", rec.offset(), previouslyCompleted);
+//        return previouslyCompleted;
+//    }
 
     /**
      * Check we have capacity in offset storage to process more messages
@@ -339,12 +311,7 @@ public class PartitionStateManager<K, V> implements ConsumerRebalanceListener {
         log.debug("Incoming {} new records...", recordsMap.count());
         for (var partition : recordsMap.partitions()) {
             RecordsAndEpoch recordsList = recordsMap.records(partition);
-            long epochOfInboundRecords = recordsList.getEpochOfPartitionAtPoll();
-            List<ConsumerRecord<K, V>> recordPollBatch = recordsList.getRecords();
-            if (!recordPollBatch.isEmpty()) {
-                // should always not be empty, but...
-                maybeRegisterNewRecordAsWork(epochOfInboundRecords, recordsList);
-            }
+            maybeRegisterNewRecordAsWork(recordsList);
         }
     }
 
@@ -353,52 +320,64 @@ public class PartitionStateManager<K, V> implements ConsumerRebalanceListener {
      */
     // todo move into PartitionState
     // todo too deep
-    private void maybeRegisterNewRecordAsWork(@NonNull Long epochOfInboundRecords, @NonNull RecordsAndEpoch recordsList) {
-        List<ConsumerRecord<K, V>> recordPollBatch = recordsList.getRecords();
+    // todo inline - shrunk
+    private void maybeRegisterNewRecordAsWork(@NonNull RecordsAndEpoch recordsAndEpoch) {
+        Long epochOfInboundRecords = recordsAndEpoch.getEpochOfPartitionAtPoll();
+        List<ConsumerRecord<K, V>> recordPollBatch = recordsAndEpoch.getRecords();
 
-        if (!recordPollBatch.isEmpty()) {
-            final Optional<ConsumerRecord<K, V>> recOpt = getFirst(recordPollBatch);
-            //noinspection OptionalGetWithoutIsPresent -- already checked not empty
-            ConsumerRecord<K, V> sampleRecord = recOpt.get(); // NOSONAR
-            long batchStartOffset = sampleRecord.offset();
+        if (recordPollBatch.isEmpty()) {
+            log.debug("Received empty poll results? {}", recordsAndEpoch);
+        } else {
+//            // check epoch is ok
+//            // todo teach PartitionState to know it's Epoch, move this into PartitionState
+//            {
+//                final Optional<ConsumerRecord<K, V>> recOpt = getFirst(recordPollBatch);
+//                //noinspection OptionalGetWithoutIsPresent -- already checked not empty
+//                ConsumerRecord<K, V> sampleRecord = recOpt.get(); // NOSONAR
+//                long batchStartOffset = sampleRecord.offset();
+//
+//                // do epochs still match? do a proactive check, but the epoch will be checked again at work completion as well
+//                var currentPartitionEpoch = getEpochOfPartitionForRecord(sampleRecord);
+//                boolean epochsDontMatch = !Objects.equals(epochOfInboundRecords, currentPartitionEpoch);
+//                if (epochsDontMatch) {
+//                    log.debug("Inbound record of work has epoch ({}) not matching currently assigned epoch for the applicable partition ({}), skipping",
+//                            epochOfInboundRecords, currentPartitionEpoch);
+//                    return;
+//                }
+//            }
 
-            // do epochs still match? do a proactive check, but the epoch will be checked again at work completion as well
-            var currentPartitionEpoch = getEpochOfPartitionForRecord(sampleRecord);
-            if (Objects.equals(epochOfInboundRecords, currentPartitionEpoch)) {
+            PartitionState<K, V> partitionState = getPartitionState(partition);
+            partitionState.maybeRegisterNewRecordsAsWork(recordsAndEpoch);
 
-                // todo move to partition state from here, as epoch apparently has to be tracked in PSM
-                if (isPartitionRemovedOrNeverAssigned(sampleRecord)) {
-                    log.debug("Record in buffer for a partition no longer assigned. Dropping. TP: {} rec: {}", toTopicPartition(sampleRecord), sampleRecord);
-                } else {
-                    //noinspection OptionalGetWithoutIsPresent -- already checked not empty
-                    long batchEndOffset = getLast(recordPollBatch).get().offset(); // NOSONAR
-
-                    TopicPartition partition = new TopicPartition(sampleRecord.topic(), sampleRecord.partition());
-                    getPartitionState(partition).maybeTruncate(batchStartOffset, batchEndOffset);
-
-                    maybeRegisterNewRecordAsWork(epochOfInboundRecords, recordPollBatch);
-                }
-            } else {
-                log.debug("Inbound record of work has epoch ({}) not matching currently assigned epoch for the applicable partition ({}), skipping",
-                        epochOfInboundRecords, currentPartitionEpoch);
-            }
+//            // todo move to partition state from here, as epoch apparently has to be tracked in PSM
+//            if (isPartitionRemovedOrNeverAssigned(sampleRecord)) {
+//                log.debug("Record in buffer for a partition no longer assigned. Dropping. TP: {} rec: {}", toTopicPartition(sampleRecord), sampleRecord);
+//            } else {
+//                //noinspection OptionalGetWithoutIsPresent -- already checked not empty
+//                long batchEndOffset = getLast(recordPollBatch).get().offset(); // NOSONAR
+//
+//                TopicPartition partition = new TopicPartition(sampleRecord.topic(), sampleRecord.partition());
+//                partitionState.maybeTruncate(batchStartOffset, batchEndOffset);
+//
+//                maybeRegisterNewRecordAsWork(epochOfInboundRecords, recordPollBatch);
+//            }
         }
     }
 
-    // todo move to partition state
-    private void maybeRegisterNewRecordAsWork(Long epochOfInboundRecords, List<ConsumerRecord<K, V>> recordPollBatch) {
-        for (var aRecord : recordPollBatch) {
-            if (isRecordPreviouslyCompleted(aRecord)) {
-                log.trace("Record previously completed, skipping. offset: {}", aRecord.offset());
-            } else {
-                //noinspection ObjectAllocationInLoop
-                var work = new WorkContainer<>(epochOfInboundRecords, aRecord, module);
-
-                sm.addWorkContainer(work);
-                addNewIncompleteWorkContainer(work);
-            }
-        }
-    }
+//    // todo move to partition state
+//    private void maybeRegisterNewRecordAsWork(Long epochOfInboundRecords, List<ConsumerRecord<K, V>> recordPollBatch) {
+//        for (var aRecord : recordPollBatch) {
+//            if (isRecordPreviouslyCompleted(aRecord)) {
+//                log.trace("Record previously completed, skipping. offset: {}", aRecord.offset());
+//            } else {
+//                //noinspection ObjectAllocationInLoop
+//                var work = new WorkContainer<>(epochOfInboundRecords, aRecord, module);
+//
+//                sm.addWorkContainer(work);
+//                addNewIncompleteWorkContainer(work);
+//            }
+//        }
+//    }
 
     public Map<TopicPartition, OffsetAndMetadata> collectDirtyCommitData() {
         var dirties = new HashMap<TopicPartition, OffsetAndMetadata>();
@@ -415,39 +394,16 @@ public class PartitionStateManager<K, V> implements ConsumerRebalanceListener {
                 .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)));
     }
 
-    public boolean couldBeTakenAsWork(WorkContainer<?, ?> workContainer) {
-        if (checkIfWorkIsStale(workContainer)) {
-            log.debug("Work is in queue with stale epoch or no longer assigned. Skipping. Shard it came from will/was removed during partition revocation. WC: {}", workContainer);
-            return false;
-        } else if (isAllowedMoreRecords(workContainer)) {
-            return true;
-        } else if (isBlockingProgress(workContainer)) {
-            // allow record to be taken, even if partition is blocked, as this record completion may reduce payload size requirement
-            return true;
-        } else {
-            log.debug("Not allowed more records for the partition ({}) as set from previous encode run (blocked), that this " +
-                            "record ({}) belongs to, due to offset encoding back pressure, is within the encoded payload already (offset lower than highest succeeded, " +
-                            "not in flight ({}), continuing on to next container in shardEntry.",
-                    workContainer.getTopicPartition(), workContainer.offset(), workContainer.isNotInFlight());
-            return false;
-        }
+    public boolean couldBeTakenAsWork(WorkContainer<K, V> workContainer) {
+        return getPartitionState(workContainer)
+                .couldBeTakenAsWork(workContainer);
     }
 
-    /**
-     * If the record is below the highest succeeded offset, then it is or will be represented in the current offset
-     * encoding.
-     * <p>
-     * This may in fact be THE message holding up the partition - so must be retried.
-     * <p>
-     * In which case - don't want to skip it.
-     * <p>
-     * Generally speaking, completing more offsets below the highest succeeded (and thus the set represented in the
-     * encoded payload), should usually reduce the payload size requirements.
-     */
-    private boolean isBlockingProgress(WorkContainer<?, ?> workContainer) {
-        var partitionState = getPartitionState(workContainer.getTopicPartition());
-        return workContainer.offset() < partitionState.getOffsetHighestSucceeded();
+    private PartitionState<K, V> getPartitionState(WorkContainer<K, V> workContainer) {
+        TopicPartition topicPartition = workContainer.getTopicPartition();
+        return getPartitionState(topicPartition);
     }
+
 
     public boolean isDirty() {
         return this.partitionStates.values().stream()
